cleanup: simplified mock controllers
