--- api-gateway/app/mock/permission_controllger.go
@@ -1,26 +1,0 @@
-package mock
-
-import "github.com/gin-gonic/gin"
-
-type PermissionControllerMock struct{}
-
-/* Controller interface implementations */
-func (m *PermissionControllerMock) GetAll(ctx *gin.Context) {
-	ctx.JSON(200, gin.H{"message": "GetAll"})
-}
-
-func (m *PermissionControllerMock) Get(ctx *gin.Context) {
-	ctx.JSON(200, gin.H{"message": "Get"})
-}
-
-func (m *PermissionControllerMock) Create(ctx *gin.Context) {
-	ctx.JSON(200, gin.H{"message": "Create"})
-}
-
-func (m *PermissionControllerMock) Update(ctx *gin.Context) {
-	ctx.JSON(200, gin.H{"message": "Update"})
-}
-
-func (m *PermissionControllerMock) Delete(ctx *gin.Context) {
-	ctx.JSON(200, gin.H{"message": "Delete"})
-}
